Refactored times and reps

+++ WorkoutTracker/SwiftUIViews/WorkoutSelectView.swift
@@ -11,13 +11,13 @@ struct Exercise {
     var name: String
     var reps: Int
     var time: Double
-    var rest: Int
+    var rest: Double
 }
 
 let Workout = [
-    Exercise(name: "Pull-up", reps: 10, time: 45.0, rest: 45),
-    Exercise(name: "Push-up", reps: 10, time: 30.0, rest: 45),
-    Exercise(name: "Sit-up", reps: 20, time: 60.0, rest: 45),
+    Exercise(name: "Pull-up", reps: 10, time: 45.0, rest: 45.0),
+    Exercise(name: "Push-up", reps: 10, time: 30.0, rest: 45.0),
+    Exercise(name: "Sit-up", reps: 20, time: 60.0, rest: 45.0),
 
 ]
